# swagger.yml
openapi: 3.0.0
info:
  title: School Management System
  description: This is a School Management System which provides automation of various academic and administrative operations.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://sch-management-server.herokuapp.com/api
    description: Production Server
paths:
...
  /teacher/fetchProfile:
    get:
      tags:
        - Teacher
      summary: Fetch Teacher Profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Teacher Profile fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Teacher Not Found
        '500':
          description: Internal Server Error
  /teacher/updateProfile:
    put:
      tags:
        - Teacher
      summary: Update Teacher Profile
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/TeacherProfile'
      responses:
        '200':
          description: Teacher Profile Updated Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /teacher/students:
    get:
      tags:
        - Teacher
      summary: Fetch all the students in sorted order
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Students Fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
...
  /teacher/fetchScorecards:
    get:
      tags:
        - Teacher
      summary: Fetch all the scorecards
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scorecard Fetched Successfully
        '401':
          description: Not Authorized to access this resource
        '500':
          description: Internal Server Error
...
  /teacher/scorecard/{scorecardId}:
    delete:
      tags:
        - Teacher
      summary: Delete Scorecard
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: scorecardId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scorecard Deleted Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Scorecard Not Found
        '500':
          description: Internal Server Error
  /teacher/rankings:
    get:
      tags:
        - Teacher
      summary: Fetch all the students according to their rankings in Exams.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Students Fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    TeacherProfileInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        dob:
          type: string
          format: date-time
        gender:
          type: string
        qualification:
          type: string
        subjectToTeach:
          type: string
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    TeacherProfile:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TeacherProfileInput'
        description: Teacher Profile
...